:art: Added type casting to sort methods

+++ src/main/java/com/lambdaschool/countries/AgeController.java
@@ -15,19 +15,19 @@ public class AgeController {
     @GetMapping(value = "/age/{age}", produces = {"application/json"})
     public ResponseEntity<?> getMedianAgeOver(@PathVariable int age) {
         ArrayList<Country> rtnCountries = CountriesApplication.ourCountryList.findCountries(c -> c.getMedianAge() > age);
-        rtnCountries.sort((o1, o2) -> o1.getName().compareToIgnoreCase(o2.getName()));
+        rtnCountries.sort((o1, o2) -> (int)(o1.getName().compareToIgnoreCase(o2.getName())));
         return new ResponseEntity<>(rtnCountries, HttpStatus.OK);
     }
 
     @GetMapping(value = "/min", produces = {"application/json"})
     public ResponseEntity<?> getMinAge() {
-        CountriesApplication.ourCountryList.countryList.sort((o1, o2) -> o1.getMedianAge() - o2.getMedianAge());
+        CountriesApplication.ourCountryList.countryList.sort((o1, o2) -> (int)(o1.getMedianAge() - o2.getMedianAge()));
         return new ResponseEntity<>(CountriesApplication.ourCountryList.countryList.get(0), HttpStatus.OK);
     }
 
     @GetMapping(value = "/max", produces = {"application/json"})
     public ResponseEntity<?> getMaxAge() {
-        CountriesApplication.ourCountryList.countryList.sort((o1, o2) -> o2.getMedianAge() - o1.getMedianAge());
+        CountriesApplication.ourCountryList.countryList.sort((o1, o2) -> (int)(o2.getMedianAge() - o1.getMedianAge()));
         return new ResponseEntity<>(CountriesApplication.ourCountryList.countryList.get(0), HttpStatus.OK);
     }
